requirements.txt
streamlit
pandas
gspread>=5.7.0
oauth2client>=4.1.3
google-auth>=2.20.0
